Revert "Merge https://github.com/tensorflow/tensorflow"
This reverts commit 842a49c590ced2877b794052d1256c263a1ea506, reversing
changes made to fa5a8d25dd7a610a58f74e9fe58b9d3a526d708c.

+++ tensorflow/python/ops/linalg/inverse_registrations.py
@@ -112,8 +112,7 @@ def _inverse_circulant(circulant_operator):
       is_non_singular=circulant_operator.is_non_singular,
       is_self_adjoint=circulant_operator.is_self_adjoint,
       is_positive_definite=circulant_operator.is_positive_definite,
-      is_square=True,
-      input_output_dtype=circulant_operator.dtype)
+      is_square=True)
 
 
 @linear_operator_algebra.RegisterInverse(
